requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.1.1
dnspython==2.3.0
frozenlist==1.3.3
fsspec==2023.1.0
gcsfs==2023.1.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.0
six==1.16.0
urllib3==1.26.14
yarl==1.8.2
